Add unit tests for SearchBar
